refactor(monitors): extract buildMonitorAlertSlackMessage into shared
