+++ packages/shared/package.json
@@ -138,6 +138,7 @@
     "oci-common": "^2.125.0",
     "oci-objectstorage": "^2.125.0",
     "safe-regex2": "^5.0.0",
+    "slackify-markdown": "^5.0.0",
     "undici": "^7.24.6",
     "uuid": "^14.0.0",
     "winston": "^3.15.0",
