+++ packages/shared/src/features/monitors/server.ts
@@ -6,3 +6,4 @@ export * from "./service";
 export * from "./scheduler/types";
 export * from "./scheduler/scheduler";
 export * from "./processor/processor";
+export * from "./processor/buildMonitorAlertSlackMessage";
